+++ features/admin.core.v1/models/config.ts
@@ -560,6 +560,10 @@ export interface UIConfigInterface extends CommonUIConfigInterface<FeatureConfig
      * Multi-tenancy related configurations.
      */
     multiTenancy: MultiTenancyConfigInterface;
+    /**
+     * Async Operation Polling Interval.
+     */
+    asyncOperationStatusPollingInterval: number;
 }
 
 /**
@@ -649,6 +653,7 @@ export interface ServiceResourceEndpointsInterface extends ClaimResourceEndpoint
     me: string;
     saml2Meta: string;
     wellKnown: string;
+    asyncStatus: string;
 }
 
 export interface ResourceEndpointsInterface {