+++ modules/i18n/src/models/namespaces/applications-ns.ts
@@ -273,6 +273,12 @@ export interface ApplicationsNS {
             content: string;
             assertionHint: string;
         };
+        inProgressReshare: {
+            header: string;
+            message: string;
+            content: string;
+            assertionHint: string;
+        };
         clientSecretHashDisclaimer: {
             modal: {
                 header: string;
@@ -771,6 +777,12 @@ export interface ApplicationsNS {
             sharedAccess: {
                 subTitle: string;
                 tabName: string;
+                sharedAccessStatusOptions: {
+                    all: string;
+                    success: string;
+                    failed: string;
+                    partiallyCompleted: string;
+                }
             };
             shareApplication: {
                 heading: string;
@@ -811,7 +823,41 @@ export interface ApplicationsNS {
                         description: string;
                     };
                 };
+                completedSharingNotification: {
+                    failure: {
+                        message: string;
+                        description: string;
+                    },
+                    success: {
+                        message: string;
+                        description: string;
+                    },
+                    partialSuccess: {
+                        message: string;
+                        description: string;
+                    }
+                },
+                addAsyncSharingNotification: {
+                    message: string,
+                    description: string,
+                },
+                asyncOperationStatus: {
+                    inProgress: {
+                        heading: string,
+                        description: string
+                    },
+                    completed: {
+                        heading: string,
+                        description: string,
+                        actionText: string
+                    }
+                },
                 switchToSelectiveShareFromSharingWithAllSuborgsWarning: string;
+                applicationShareFailureSummaryDefaultStatus: {
+                    success: string,
+                    failed: string,
+                    partiallyCompleted: string
+                }
             };
             apiAuthorization: {
                 m2mPolicyMessage: string;
@@ -2960,5 +3006,14 @@ export interface ApplicationsNS {
                 description2: string;
             };
         };
+        sharedAccessStatus: {
+            heading: string,
+            subHeading: string,
+            actionText: string,
+            banner: {
+                partiallyCompleted: string,
+                failed: string
+            }
+        }
     };
 }